+++ components/design-system/card/Card.tsx
@@ -12,12 +12,13 @@ type CardProps = {
 };
 
 export const useCardStyles = () => {
-  const { boxShadow, radii, colors, spacing } = useTheme();
+  const { boxShadow, radii, colors, spacing, maxWidth } = useTheme();
   return {
     ...boxShadow[1],
     background: colors.background.emphasis,
     padding: spacing.large,
-    borderRadius: radii.large
+    borderRadius: radii.large,
+    maxWidth: maxWidth.content
   };
 };
